refactor(benchmarks): extract route surface hierarchy
